refactor(credentials): move module-specific CredentialReaderOptions to each module
Core should only contain generic credential options (ForPin, ForPuk,
ForPassphrase, ForHexKey). Module-specific presets belong in the modules
that own the domain knowledge:

- FidoCredentialOptions.ForFido2Pin() in Yubico.YubiKit.Fido2
- OpenPgpCredentialOptions.ForOpenPgpPin/AdminPin/ResetCode() in OpenPgp
- OathCredentialOptions.ForOathPassword() in Oath

This keeps Core free of FIDO2/OpenPGP/OATH domain knowledge and follows
the existing dependency direction (module → Core, never reverse).
